Add non-stereo channel support
This does not change the API, but fills in wChannelConfig with a
better value for the number of common input channels.

Without it, host audio APIs can fail to capture or render for
>2 channels.

A more comprehensive update would require an API change that
sets channel config by a bitmask, e.g. for a 7-channel setup and
such.